refactor: Separa app.cjs em rotas, controlles e services. Prepara a arquitetura para JWT
