+++ package.json
@@ -12,7 +12,7 @@
     "body-parser": "^2.2.0",
     "cors": "^2.8.5",
     "express": "^5.1.0",
-    "nodemon": "^3.1.10",
+    "nodemon": "^3.1.11",
     "readline": "^1.3.0"
   },
   "devDependencies": {
